p2phttp: fix stalling git-annex get
A race condition caused a small fraction of requests to hang with the
object mostly transferred. Or, in some cases, STM deadlock messages to
be displayed without a hang.

See comments for analysis. I don't entirely understand what is going on
with all the filling of endv, but this clearly fixes the race.