S3: Remote can be configured with an x-amz-tagging header.
Needs aws-0.25, which was just released.

stack.yaml: Use aws-0.25.

This risks some foot shooting if the user configures a bucket to hide
objects that have a given tag, and sets up the S3 special remote with it,
but then uses a build of git-annex that does not support sending it.

I considered making git-annex error out if it's built with too old a
version of aws. But that leaves the problem of older versions of git-annex,
which will just ignore the x-amz-tagging config. So it didn't seem worth
the bother to have a new failure mode. The user of this will just need to
be careful of their versions if they are using it in a security context.

The use of parseQueryText is kind of silly, because aws just reassembles it
back to a query string. But it made sense for the aws interface to be
key/value pairs. And it seems to make sense for the git-annex config
interface to be a query string. At least, I could not think of a better way
to handle the config.

Sponsored-by: Dartmouth College's OpenNeuro project