Added TRANSFER-RETRIEVE-URL extension to the external special remote protocol
Since retrieveKeyFileM has to use fileRetriever before it sees this
response, which uses tailVerify, it's unfortunately not possible to
stream the url download to do incremental verification. That would be
more efficient.

Similarly, watchFileSize does some extra work, but the progress meter is
updated as the content streams in.

The downloadFailed case is never reached I think, since the url list
provided to downloadUrl' is not empty.

Sponsored-by: Dartmouth College's OpenNeuro project