--presentsince, --lackingsince, and --changedsince
There are some complications around vector clocks, which basically come
down to, when the clocks of a repository has been wrong, these won't
reflect actual real times. Which I think is too obvious to document to the
user.

The possibility of a distributed system updating the location log
unncessarily does seem like something a user might get confused by. It
would be a fairly rare situation that caused it though. For example
`git-annex fsck --from remote` will update the location logs for changes
it finds, and when the git-annex branch is out of sync with the remote,
will make unncessary updates. Once those get synced to the remote,
in that repository eg --presentsince=here:interval will only see the
most recent log entry, which is at a later point in time than when
the content was actually present in the repository. It would be good to
document this perhaps, but it's a rather complicated and unusual situation.